fixed some issues with wifiscanner: Location was not saved to the database, so wifiscanresult was missing the result.
Update issue #9
added WifiScanResult to MeasuringPoint

git-svn-id: https://wificompass.googlecode.com/svn/trunk@61 6f3e661d-c991-e48d-a97c-fec7ba324fcc

+++ src/at/fhstp/wificompass/userlocation/LocationProvider.java
@@ -36,19 +36,19 @@ public interface LocationProvider {
 	  * get X coordinate
 	 * @return x coordinate
 	 */
-	double getLocationX();
+	float getLocationX();
 	 
 	 /**
 	  * get Y coordinate
 	 * @return y coordinate
 	 */
-	double getLocationY();
+	float getLocationY();
 
 	/**
 	 * get Z coordinate if applicable, otherwise zero
 	 * @return z coordinate
 	 */
-	double getLocationZ();
+	float getLocationZ();
 	
 	/**
 	 * get current Location of the user
